refactor: reduce clutter, extract classes, shared conditions
