+++ src/superLabeler.ts
@@ -1,8 +1,9 @@
 import fs from 'fs'
 import * as github from '@actions/github'
 import { GitHub } from '@actions/github'
-import { Config, Options, PRContext, IssueContext } from './types'
-import { labelHandler, contextHandler } from './utils'
+import { CurContext, Config, Options, labelIdToName } from './types'
+import { labelHandler } from './labelHandler'
+import { contextHandler } from './contextHandler'
 import { log } from './'
 
 let local: any
@@ -26,6 +27,10 @@ try {
 export default class SuperLabeler {
   client: GitHub
   opts: Options
+  configJSON: Options['configJSON']
+  configPath: Options['configPath']
+  dryRun: Options['dryRun']
+  repo = context.repo || {}
 
   /**
    * @author IvanFon, TGTGamer, jbinda
@@ -35,6 +40,9 @@ export default class SuperLabeler {
     log(`Superlabeller Constructed: ${options}`, 1)
     this.client = client
     this.opts = options
+    this.configJSON = options.configJSON
+    this.configPath = options.configPath
+    this.dryRun = options.dryRun
   }
 
   /**
@@ -43,153 +51,217 @@ export default class SuperLabeler {
    * @since 1.0.0
    */
   async run() {
-    try {
-      const configJSON = this.opts.configJSON
-      const configPath = this.opts.configPath
-      const dryRun = this.opts.dryRun
-      const repo = context.repo || {}
-      if (dryRun) repo.repo = process.env.GITHUB_REPOSITORY || 'Unknown'
-      if (dryRun) repo.owner = process.env.GITHUB_REPOSITORY_OWNER || 'Unknown'
+    if (this.dryRun) this.repo.repo = process.env.GITHUB_REPOSITORY || 'Unknown'
+    if (this.dryRun)
+      this.repo.owner = process.env.GITHUB_REPOSITORY_OWNER || 'Unknown'
+    log(`Repo data: ${this.repo.owner}/${this.repo.repo}`, 1)
 
-      log(`Repo data: ${repo.owner}/${repo.repo}`, 1)
+    /**
+     * Capture and log context to debug for Local Running
+     * @author TGTGamer
+     * @since 1.0.0
+     */
+    log(
+      `Context for local running. See readme.md for information on how to setup local running: ${JSON.stringify(
+        context
+      )}`,
+      1
+    )
 
-      /**
-       * Capture and log context to debug for Local Running
-       * @author TGTGamer
-       * @since 1.0.0
-       */
+    /**
+     * Process the config
+     * @author TGTGamer
+     * @since 1.1.0
+     */
+    const configs = await this.processConfig().catch(err => {
+      log(`Error thrown while processing config: ` + err, 5)
+      throw err
+    })
+    log(`Config: ${JSON.stringify(configs)}`, 1)
 
-      log(
-        `Context for local running. See readme.md for information on how to setup local running: ${JSON.stringify(
-          context
-        )}`,
-        1
-      )
+    /**
+     * Get the context
+     * @author TGTGamer
+     * @since 1.1.0
+     */
+    const curContext = await this.processContext().catch(err => {
+      log(`Error thrown while processing context: ` + err, 5)
+      throw err
+    })
 
-      /**
-       * Get the configuration
-       * @author IvanFon, TGTGamer, jbinda
-       * @since 1.0.0
-       */
-      let config: Config
-      if (!configJSON) {
-        if (!fs.existsSync(configPath)) {
-          throw new Error(`config not found at "${configPath}"`)
-        }
-        config = await JSON.parse(fs.readFileSync(configPath).toString())
-        log(`Config: ${JSON.stringify(config)}`, 1)
-      } else {
-        config = configJSON
+    /**
+     * Combine the Shared & Context.type Configs
+     * @author TGTGamer
+     * @since 1.1.0
+     */
+    for (const config in configs.shared) {
+      if (!configs[curContext.type][config]) {
+        configs[curContext.type][config] = configs.shared[config]
       }
+    }
 
+    /**
+     * Syncronise the labels
+     * @author TGTGamer
+     * @since 1.1.0
+     */
+    await this.syncLabels(configs).catch(err => {
+      log(`Error thrown while syncronising labels: ` + err, 5)
+      throw err
+    })
+
+    /**
+     * Convert label ID's to Names
+     * @author TGTGamer
+     * @since 1.1.0
+     */
+    const labelIdToName: labelIdToName = await Object.entries(
+      configs.labels
+    ).reduce((acc: { [key: string]: string }, cur) => {
+      acc[cur[0]] = cur[1].name
+      return acc
+    }, {})
+
+    /**
+     * Apply the context
+     * @author TGTGamer
+     * @since 1.1.0
+     */
+    await this.applyContext(curContext, configs, labelIdToName).catch(err => {
+      log(`Error thrown while applying context: ` + err, 5)
+      throw err
+    })
+  }
+
+  /**
+   * Get the configuration
+   * @author IvanFon, TGTGamer, jbinda
+   * @since 1.0.0
+   */
+  async processConfig(): Promise<Config> {
+    if (!this.configJSON) {
+      if (!fs.existsSync(this.configPath)) {
+        throw new Error(`config not found at "${this.configPath}"`)
+      }
+      return await JSON.parse(fs.readFileSync(this.configPath).toString())
+    } else {
+      return this.configJSON
+    }
+  }
+
+  /**
+   * Handle the context
+   * @author IvanFon, TGTGamer, jbinda
+   * @since 1.0.0
+   */
+  async processContext() {
+    let curContext: CurContext
+
+    if (context.payload.pull_request) {
       /**
-       * Handle the context
+       * Pull Request Context
        * @author IvanFon, TGTGamer, jbinda
        * @since 1.0.0
        */
-      let curContext:
-        | { type: 'pr'; context: PRContext }
-        | { type: 'issue'; context: IssueContext }
-
-      if (context.payload.pull_request) {
-        const ctx = await contextHandler
-          .parsePR(context, this.client, repo)
-          .catch(err => {
-            log(`Error thrown while parsing PR context: ` + err, 5)
-            throw err
-          })
-        if (!ctx) {
-          throw new Error('Pull Request not found on context')
-        }
-        log(`PR context: ${JSON.stringify(ctx)}`, 1)
-        curContext = {
-          type: 'pr',
-          context: ctx
-        }
-      } else if (context.payload.issue) {
-        const ctx = await contextHandler.parseIssue(context).catch(err => {
-          log(`Error thrown while parsing issue context: ` + err, 5)
+      const ctx = await contextHandler
+        .parsePR(context, this.client, this.repo)
+        .catch(err => {
+          log(`Error thrown while parsing PR context: ` + err, 5)
           throw err
         })
-        if (!ctx) {
-          throw new Error('Issue not found on context')
-        }
-        log(`issue context: ${JSON.stringify(ctx)}`, 1)
-
-        curContext = {
-          type: 'issue',
-          context: ctx
-        }
-      } else {
-        log(
-          `There is no context to parse: ${JSON.stringify(context.payload)}`,
-          3
-        )
-        throw new Error('There is no context')
+      if (!ctx) {
+        throw new Error('Pull Request not found on context')
       }
-
+      log(`PR context: ${JSON.stringify(ctx)}`, 1)
+      curContext = {
+        type: 'pr',
+        context: ctx
+      }
+    } else if (context.payload.issue) {
       /**
-       * Syncronise labels to repository
+       * Issue Context
        * @author IvanFon, TGTGamer, jbinda
        * @since 1.0.0
        */
+      const ctx = await contextHandler.parseIssue(context).catch(err => {
+        log(`Error thrown while parsing issue context: ` + err, 5)
+        throw err
+      })
+      if (!ctx) {
+        throw new Error('Issue not found on context')
+      }
+      log(`issue context: ${JSON.stringify(ctx)}`, 1)
+
+      curContext = {
+        type: 'issue',
+        context: ctx
+      }
+    } else {
+      /**
+       * No Context
+       * @author TGTGamer
+       * @since 1.1.0
+       */
+      log(`There is no context to parse: ${JSON.stringify(context.payload)}`, 3)
+      throw new Error('There is no context')
+    }
+    return curContext
+  }
+
+  /**
+   * Syncronise labels to repository
+   * @author IvanFon, TGTGamer, jbinda
+   * @since 1.0.0
+   */
+  async syncLabels(config: Config) {
+    await labelHandler
+      .syncLabels({
+        client: this.client,
+        repo: this.repo,
+        config: config.labels,
+        dryRun: this.dryRun
+      })
+      .catch((err: { message: string | Error }) => {
+        log(`Error thrown while handling syncLabels tasks: ${err.message}`, 5)
+      })
+  }
+
+  /**
+   * Apply labels to context
+   * @author IvanFon, TGTGamer, jbinda
+   * @since 1.0.0
+   */
+  async applyContext(
+    curContext: CurContext,
+    config: Config,
+    labelIdToName: labelIdToName
+  ) {
+    if (curContext.type === 'pr') {
       await labelHandler
-        .syncLabels({
+        .applyPR({
           client: this.client,
-          repo,
-          config: config.labels,
-          dryRun
+          config: config.pr,
+          labelIdToName,
+          prContext: curContext.context,
+          repo: this.repo,
+          dryRun: this.dryRun
         })
         .catch((err: { message: string | Error }) => {
-          log(`Error thrown while handling syncLabels tasks: ${err.message}`, 5)
+          log(`Error thrown while handling PRLabel tasks: ${err.message}`, 5)
         })
-
-      // Mapping of label ids to Github names
-      const labelIdToName = await Object.entries(config.labels).reduce(
-        (acc: { [key: string]: string }, cur) => {
-          acc[cur[0]] = cur[1].name
-          return acc
-        },
-        {}
-      )
-
-      /**
-       * Apply labels to context
-       * @author IvanFon, TGTGamer, jbinda
-       * @since 1.0.0
-       */
-      if (curContext.type === 'pr') {
-        await labelHandler
-          .applyPR({
-            client: this.client,
-            config: config.pr,
-            labelIdToName,
-            prContext: curContext.context,
-            repo,
-            dryRun
-          })
-          .catch((err: { message: string | Error }) => {
-            log(`Error thrown while handling PRLabel tasks: ${err.message}`, 5)
-          })
-      } else if (curContext.type === 'issue') {
-        await labelHandler
-          .applyIssue({
-            client: this.client,
-            config: config.issue,
-            issueContext: curContext.context,
-            labelIdToName,
-            repo,
-            dryRun
-          })
-          .catch((err: { message: string | Error }) => {
-            log(
-              `Error thrown while handling issueLabel tasks: ${err.message}`,
-              5
-            )
-          })
-      }
-    } catch (err) {
-      log(`Error in supper labeler: ` + err.message, 5)
+    } else if (curContext.type === 'issue') {
+      await labelHandler
+        .applyIssue({
+          client: this.client,
+          config: config.issue,
+          issueContext: curContext.context,
+          labelIdToName,
+          repo: this.repo,
+          dryRun: this.dryRun
+        })
+        .catch((err: { message: string | Error }) => {
+          log(`Error thrown while handling issueLabel tasks: ${err.message}`, 5)
+        })
     }
   }
 }
